Restructure mathematical representations.
