+++ include/Mathematical/Representation/ChemicalRankFourTensor.hpp
@@ -18,8 +18,8 @@
 #pragma once
 
 
-#include "Mathematical/SquareMatrix.hpp"
-#include "Mathematical/SquareRankFourTensor.hpp"
+#include "Mathematical/Representation/SquareMatrix.hpp"
+#include "Mathematical/Representation/SquareRankFourTensor.hpp"
 
 #include <iostream>
