+++ include/Mathematical/Representation/Tensor.hpp
@@ -18,7 +18,7 @@
 #pragma once
 
 
-#include "Mathematical/Matrix.hpp"
+#include "Mathematical/Representation/Matrix.hpp"
 #include "typedefs.hpp"
 
 #include <unsupported/Eigen/CXX11/Tensor>
