+++ include/RDM/TwoRDM.hpp
@@ -18,7 +18,7 @@
 #pragma once
 
 
-#include "Mathematical/ChemicalRankFourTensor.hpp"
+#include "Mathematical/Representation/ChemicalRankFourTensor.hpp"
 #include "RDM/OneRDM.hpp"
