+++ src/Mathematical/Optimization/NewtonMinimizer.cpp
@@ -18,7 +18,7 @@
 #include "Mathematical/Optimization/NewtonMinimizer.hpp"
 
 #include "Mathematical/Optimization/NewtonSystemOfEquationsSolver.hpp"
-#include "Mathematical/SquareMatrix.hpp"
+#include "Mathematical/Representation/SquareMatrix.hpp"
 #include "typedefs.hpp"
 
 #include <iostream>
