+++ src/Mathematical/Optimization/step.cpp
@@ -17,7 +17,7 @@
 // 
 #include "Mathematical/Optimization/step.hpp"
 
-#include "Mathematical/SquareMatrix.hpp"
+#include "Mathematical/Representation/SquareMatrix.hpp"
 
 
 namespace GQCP {
