+++ tests/Mathematical/Optimization/NewtonMinimizer_test.cpp
@@ -20,7 +20,7 @@
 #include <boost/test/unit_test.hpp>
 
 #include "Mathematical/Optimization/NewtonMinimizer.hpp"
-#include "Mathematical/SquareMatrix.hpp"
+#include "Mathematical/Representation/SquareMatrix.hpp"
 
 
 /*
